# docs/openapi.yaml
openapi: 3.0.3
info:
  title: HappYness Chat API
  description: |
    Real-time chat backend for the HappYness platform. Exposes REST endpoints for
    chat/message/read-marker management and a per-user WebSocket endpoint for live
    delivery. JWT (HS512) auth is required for `/api/**` REST routes and for the
    WebSocket endpoint.

    See `WebSocket protocol` at the bottom of this document for the per-user `/api/ws`
    event envelope.
  version: "1.0.0"
  contact:
    name: HappYness Project
    url: https://github.com/HappYness-Project/chatApi
servers:
  - url: http://localhost:8070
    description: Local development (docker-compose)
  - url: https://{host}
    description: Deployed environment
    variables:
      host:
        default: chats.example.com
...
security:
  - bearerAuth: []
paths:
...
  /api/chats/{chatID}/messages:
    parameters:
      - $ref: "#/components/parameters/ChatIDPath"
      - in: query
        name: limit
        schema: { type: integer, minimum: 1, default: 120 }
        description: Maximum number of messages to return.
      - in: query
        name: offset
        schema: { type: integer, minimum: 0, default: 0 }
        description: Number of messages to skip for pagination.
    get:
      tags: [Messages]
      summary: List messages for a chat
      responses:
        "200":
          description: Messages page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessagesPage"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        HS512 JWT issued by the platform auth service. The token's `nameid` (or
        `sub`) claim must be a UUID identifying the user. Required for all
        `/api/**` REST routes. WebSocket endpoints accept the token as
        `?token=` instead of an `Authorization` header.
  parameters:
    ChatIDPath:
      in: path
      name: chatID
      required: true
      schema: { type: string, format: uuid }
      description: Chat UUID (v7).
...
  responses:
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
...
    InternalError:
      description: Unexpected server error
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/ProblemDetails" }
  schemas:
...
    ProblemDetails:
      type: object
      description: RFC 7807-style problem details (Content-Type `application/problem+json`).
      properties:
        type: { type: string }
        title: { type: string }
        error_code: { type: string }
        detail: { type: string }
        instance: { type: string }
      required: [title, error_code]
...
    Message:
      type: object
      properties:
        id: { type: string, format: uuid }
        chat_id: { type: string, format: uuid }
        sender_id: { type: string, format: uuid }
        content: { type: string }
        message_type: { type: string, example: text }
        created_at: { type: string, format: date-time }
        read_status: { type: boolean }
        deleted_at:
          type: string
          format: date-time
          nullable: true
        deleted_by:
          type: string
          format: uuid
          nullable: true
      required: [id, chat_id, sender_id, content, message_type, created_at, read_status]
    MessagesPage:
      type: object
      properties:
        messages:
          type: array
          items: { $ref: "#/components/schemas/Message" }
        count: { type: integer }
      required: [messages, count]
...